backend/requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.3

# Environment variables
python-dotenv==1.0.0

# AI/LLM integration
openai>=1.0.0

# Date handling
python-dateutil>=2.8.2
